Use node-compatible http library, axios
goog.net.XhrIO only works in the context of a browser with access to the XMLHttpRequest
api, which doesn't exist in node. Our workaround has been to require users to require
another library that provides this functionality:

global.XMLHttpRequest = require("xmlhttprequest").XMLHttpRequest;

This changes our library to remove that external dependency.